+++ demo/src/main.ts
@@ -8,5 +8,16 @@ if (environment.production) {
   enableProdMode();
 }
 
-platformBrowserDynamic().bootstrapModule(AppModule)
-  .catch(err => console.log(err));
+platformBrowserDynamic().bootstrapModule(AppModule).then(
+  (ref) => {
+    if (window['ngRef']) {
+      window['ngRef'].destroy();
+    }
+
+    window['ngRef'] = ref;
+  }
+).catch(
+  (err) => {
+    console.log(err)
+  }
+);