+++ src/auto-complete-styles.model.ts
@@ -1,0 +1,5 @@
+export class AutoCompleteStyles {
+  public list = {};
+  public listItem = {};
+  public searchbar = {};
+}