+++ CHANGELOG.md
@@ -7,6 +7,106 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [Released]
+
+## [2.5.0]
+### Added
+* Ability to pass custom styles on elements within the Shadow DOM by [HashNotAdam](https://github.com/HashNotAdam)
+
+## [2.4.2]
+### Fixed
+- Show loading when dataProvider is a function by [jrquick17](https://github.com/jrquick17)
+- Do not let loading spinner and clear icon overlap by [jrquick17](https://github.com/jrquick17)
+
+## [2.4.1]
+### Fixed
+- Keyup events when `[useIonInput]="true"` by [jrquick17](https://github.com/jrquick17)
+
+## [2.4.0]
+### Added
+- `itemsCleared` event by [jrquick17](https://github.com/jrquick17)
+- Hide menu list when user clicks escape by [jrquick17](https://github.com/jrquick17)
+- Navigate options using arrow keys by [jrquick17](https://github.com/jrquick17)
+- Select item using enter key by [jrquick17](https://github.com/jrquick17)
+
+### Removed
+- Demo from npm installs by [jrquick17](https://github.com/jrquick17)
+
+### Moved
+- Use stylesheet within component by [jrquick17](https://github.com/jrquick17)
+
+## [2.3.2]
+### Fixed
+- Event propagation when item is selected by [jrquick17](https://github.com/jrquick17)
+
+## [2.3.1]
+### Fixed
+- getLabel function in selection template by [jrquick17](https://github.com/jrquick17)
+
+## [2.3.0]
+### Added
+- `clearInvalidInput` automatically clear the input if a user does not select an item by [jrquick17](https://github.com/jrquick17)
+- `maxSelected` allows setting the maximum allowed number of items that can be selected by [jrquick17](https://github.com/jrquick17)
+- Custom templates for when there are no matching suggestions and for displaying selected items within multi select by [jrquick17](https://github.com/jrquick17)
+
+## [2.2.1]
+### Added
+- `maxResults` to allow limiting number of suggestions by [jrquick17](https://github.com/jrquick17)
+- `build:demo` command to make testing easier by [jrquick17](https://github.com/jrquick17)
+- `enableBrowserAutocomplete` to allow enabling/disabling browser autocomplete by [jrquick17](https://github.com/jrquick17)
+
+### Changed
+- Allow exclude to be objects with labels or just the label directly by [jrquick17](https://github.com/jrquick17)
+
+### Fixed
+- Automatically remove item from array if not multi by [jrquick17](https://github.com/jrquick17)
+
+### Update
+- NPM scripts by [jrquick17](https://github.com/jrquick17)
+
+## [2.2.0]
+### Added
+- `maxResults` to allow limiting number of suggestions by [jrquick17](https://github.com/jrquick17)
+- `build:demo` command to make testing easier by [jrquick17](https://github.com/jrquick17)
+
+### Updated
+- Included console logs in the demo for testing `Output()` events by [jrquick17](https://github.com/jrquick17)
+- Deprecated fields for IonSearchBar's `showCancelButton` by [jrquick17](https://github.com/jrquick17)
+
+### Fixed
+- Build command updated to include assets by [jrquick17](https://github.com/jrquick17)
+
+## [2.1.5]
+### Removed
+- Removed deprecated by [gnowland](https://github.com/gnowland)
+
+## [2.1.4]
+### Fixed
+- Multi select getting cleared on search by [jrquick17](https://github.com/jrquick17)
+
+### Removed
+- Removed deprecated `HttpClient` by [gnowland](https://github.com/gnowland)
+
+## [2.1.3]
+### Removed
+- Remove deprecated `@angular/http` by [gnowland](https://github.com/gnowland)
+
+## [2.1.2]
+### Fixed
+- Two-way binding for [(model)] by [jrquick17](https://github.com/jrquick17)
+
+## [2.1.1]
+### Added
+- Templates for reporting issues and adding pull requests by [jrquick17](https://github.com/jrquick17)
+- Custom template demo by [jrquick17](https://github.com/jrquick17)
+- Add routing to the demo by [jrquick17](https://github.com/jrquick17)
+
+## [2.1.0]
+### Changed
+- Use ng-packagr to handle building the npm dist by [jrquick17](https://github.com/jrquick17)
+
+### Removed
+- Gulp by [jrquick17](https://github.com/jrquick17)
+
 ## [2.0.0] 
 ### Changed
 - Upgraded to support Angular 8 by [jrquick17](https://github.com/jrquick17)
@@ -15,6 +115,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added `frontIcon` to selection chips by [jrquick17](https://github.com/jrquick17)
 - Added eager loading option by [jrquick17](https://github.com/jrquick17)
+
 ### Changed
 - Use `ion-chip` for selection instead of `ion-button` by [jrquick17](https://github.com/jrquick17)
 - Made `removeIcon` optional within selection by [jrquick17](https://github.com/jrquick17)
@@ -22,8 +123,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.8.1] 
 ### Added
 - Added interface for data provider by [jrquick17](https://github.com/jrquick17)
+
 ### Changed
 - Updated dependencies by [jrquick17](https://github.com/jrquick17)
+
 ### Fixed
 - Fixed error with ionic production builds using AOT by [jrquick17](https://github.com/jrquick17)
 
@@ -31,6 +134,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added color options by [jrquick17](https://github.com/jrquick17)
 - Added multi select to the demo by [jrquick17](https://github.com/jrquick17)
+
 ### Changed
 - Updated dependencies by [jrquick17](https://github.com/jrquick17)
 - Updated to Gulp 4 by [jrquick17](https://github.com/jrquick17)
@@ -38,32 +142,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed first selectedItem not getting changed initially by [jrquick17](https://github.com/jrquick17)
 
-## [1.7.0] - 2019-05-30
+## [1.7.0]
 ### Added
 - Add exclude by [jrquick17](https://github.com/jrquick17)
 
-## [1.6.1] - 2019-05-24
+## [1.6.1]
 ### Changed
 - Updated dependencies by [jrquick17](https://github.com/jrquick17)
 
-## [1.6.1] - 2019-03-28
 ### Added
 - Add disable option by [jrquick17](https://github.com/jrquick17)
 
-## [1.6.0] - 2019-03-01
+## [1.6.0]
 ### Added
 - Support for Angular 7.0 by [jrquick17](https://github.com/jrquick17)
 - Support for Ionic 4 by [jrquick17](https://github.com/jrquick17)
 
-## [1.4.3] - 2017-05-15
+## [1.4.3]
 ### Added
 - Support for Angular 5.0
 
-## [1.4.3-beta] - 2017-05-12
+## [1.4.3-beta]
 ### Added
 - Support for ngModel
 
-## [1.4.2] - 2017-05-08
+## [1.4.2] 
 ### Added
 - Added `setFocus()` method
 - Added `getSelection()` method
@@ -72,7 +175,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Minor text changed by [kadoshms](https://github.com/kadoshms).
 
-## [1.4.1] - 2017-04-24
+## [1.4.1]
 ### Added
 - Support for rxjs subject
 
@@ -80,7 +183,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Document click handler issue
 - Removed redundant script by [kadoshms](https://github.com/kadoshms).
 
-## [1.3.0] - 2016-10-11
+## [1.3.0]
 ### Added
 - New custom-templates mechanism (please refer to the README file)
 - Added [showResultsFirst] option - calls `getItems()` when the component is tapped